# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus Security API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/tickets:
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create a support ticket (from contact form)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
  /my/tickets/{ticketId}:
    get:
      operationId: getMyTicket
      tags: [support]
      summary: Get one of the current user's tickets with its messages
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketThread"
  /my/tickets/{ticketId}/messages:
    post:
      operationId: postMyTicketMessage
      tags: [support]
      summary: Add a message to one of the current user's tickets
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketMessageInput"
      responses:
        "201":
          description: Message added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
...
components:
  schemas:
...
    CreateTicketInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        websiteUrl:
          type: ["string", "null"]
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
    SupportTicket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        status:
          type: string
        priority:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - subject
        - status
        - priority
        - createdAt
        - updatedAt
...
    TicketMessage:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        senderRole:
          type: string
        senderName:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - senderRole
        - body
        - createdAt
    TicketMessageInput:
      type: object
      properties:
        body:
          type: string
      required:
        - body
    TicketThread:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/SupportTicket"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
      required:
        - ticket
        - messages
...