# lib/api-zod/src/generated/types/supportTicket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Security API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SupportTicket {
  id: string;
  /** @nullable */
  userId?: number | null;
  name: string;
  email: string;
  subject: string;
  status: string;
  priority: string;
  createdAt: string;
  updatedAt: string;
}